chore: add utils tests
